Move test to run under linux_bpf
